refactor: Split utility functions by feature
