+++ src/utils/media.ts
@@ -1,0 +1,4 @@
+export const playAudio = (path: string) => {
+  const audio = new Audio(path);
+  audio.play();
+};
