+++ src/utils/user-agent.ts
@@ -1,0 +1,5 @@
+export const isMobile = () => {
+  const agent = navigator.userAgent;
+  const regex = /Android|iPad|iPhone/;
+  return regex.test(agent);
+};
